Fix actor switcher to respect actor mode and capabilities
- Move settings registration from class-settings.php to class-options.php
  so settings are available outside wp-admin via init hook
- Add schema with enum values for activitypub_actor_mode setting
- Update actor switcher to check user capabilities via canUser on
  ap_extrafield and ap_extrafield_blog post types
- Add capability mappings to extra field post types (activitypub for
  user fields, manage_options for blog fields)
- Auto-correct active actor when mode changes make current actor unavailable
- Link to appropriate settings page based on available actor
